requirements.txt
alembic
attrs
bcrypt
blinker
certifi
click
distlib
filelock
flasgger
Flask
Flask-Bcrypt
Flask-CORS
Flask-JWT-Extended
Flask-Mail
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
greenlet
gunicorn
importlib_metadata
importlib_resources
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
Mako
MarkupSafe
marshmallow
mistune
packaging
pipenv
pkgutil_resolve_name
platformdirs
psycopg2-binary
PyJWT
python-dotenv
PyYAML
referencing
rpds-py
six
SQLAlchemy
sqlalchemy-serializer
typing_extensions
virtualenv
Werkzeug
zipp
email_validator
